Adds local storage of recycle day
The recycle day is now being stored, and at the beginning of the app, the code now checks for an existing day that is in the future.  If the saved recycle day is still in the future, it just shows that day instead of making another API call.  This parsing and comparison was done with Moment.JS, so we now have the ability to refer to the time as something other than a string, such as changing the format of the date, or using relative time (like "in two days").